feat: update cell attribute model to support RGB, indexed colors, and inverse flags
- Replace `Int`-based cell attributes with a more structured `AttributeColor` model.
- Expand `Attributes` to support default/indexed/RGB colors and inverse/reverse-video flag.
- Refactor related APIs to use `Long` for packed attributes, ensuring future scalability.
- Update all tests and logic to align with the new attribute model and validation rules.
- Modify `AttributeCodec` for packing/unpacking long attributes, including foreground/background colors, styles, and protection flags.

+++ terminal-core/src/main/kotlin/core/api/TerminalLineApi.kt
@@ -25,11 +25,11 @@ interface TerminalLineApi {
     fun getCodepoint(col: Int): Int
 
     /**
-     * Returns the packed attribute integer for the cell at [col].
+     * Returns the packed attribute word for the cell at [col].
      * Decode with [com.gagik.core.codec.AttributeCodec] to access individual fields.
      * Use this method in production render loops; it is zero-allocation.
      */
-    fun getPackedAttr(col: Int): Int
+    fun getPackedAttr(col: Int): Long
 
     /**
      * Returns `true` if the cell at [col] holds a multi-codepoint grapheme cluster
